+++ pyproject.toml
@@ -140,4 +140,4 @@ where = ["src"]
 skip = [".github/*"]
 
 [project.entry-points.'nomad.plugin']
-export_entries_action_entry_point = "nomad_ml_workflows.actions:export_entries_action_entry_point"
+export_entries_action_entry_point = "nomad_ml_workflows.actions:export_entries"
