+++ src/nomad_ml_workflows/actions/export_entries/__init__.py
@@ -21,14 +21,16 @@ class ExportEntriesActionEntryPoint(ActionEntryPoint):
     def load(self):
         from nomad.actions import Action
 
-        from nomad_ml_workflows.actions.entries.activities import (
+        from nomad_ml_workflows.actions.export_entries.activities import (
             cleanup_artifacts,
             create_artifact_subdirectory,
             export_dataset_to_upload,
             merge_output_files,
             search,
         )
-        from nomad_ml_workflows.actions.entries.workflows import ExportEntriesWorkflow
+        from nomad_ml_workflows.actions.export_entries.workflows import (
+            ExportEntriesWorkflow,
+        )
 
         return Action(
             task_queue=self.task_queue,
@@ -43,7 +45,7 @@ class ExportEntriesActionEntryPoint(ActionEntryPoint):
         )
 
 
-export_entries_action_entry_point = ExportEntriesActionEntryPoint(
+export_entries = ExportEntriesActionEntryPoint(
     name='Export Entries Action',
     description='An action to search entries and export them as a zip file in the '
     'specified upload.',
